+++ modules/resource-metadata/main.tf
@@ -33,7 +33,7 @@ module "eventbridge" {
     crons = [
       {
         name  = "cron-for-lambda"
-        arn   = var.layer_arn == "" ? module.lambda.lambda_function_arn : module.lambdaSM.lambda_function_arn
+        arn   = var.secret_manager_enabled == false ? module.lambda.lambda_function_arn : module.lambdaSM.lambda_function_arn
         input = jsonencode({ "job" : "cron-by-rate" })
       }
     ]
@@ -54,7 +54,7 @@ resource "null_resource" "s3_bucket" {
 }
 
 module "lambda" {
-  create                 = var.layer_arn == "" ? true : false
+  create                 = var.secret_manager_enabled == false ? true : false
   depends_on             = [ null_resource.s3_bucket ]
   source                 = "terraform-aws-modules/lambda/aws"
   version                = "3.2.1"
@@ -115,7 +115,7 @@ module "lambda" {
 }
 
 module "lambdaSM" {
-  create                 = var.layer_arn != "" ? true : false
+  create                 = var.secret_manager_enabled ? true : false
   depends_on             = [ null_resource.s3_bucket ]
   source                 = "terraform-aws-modules/lambda/aws"
   version                = "3.2.1"
@@ -194,14 +194,14 @@ resource "aws_sns_topic" "this" {
 }
 
 resource "aws_secretsmanager_secret" "private_key_secret" {
-  count       = var.layer_arn != "" && var.create_secret == "True"  ? 1 : 0
+  count       = var.secret_manager_enabled && var.create_secret == "True"  ? 1 : 0
   depends_on  = [module.lambdaSM]
   name        = "lambda/coralogix/${data.aws_region.this.name}/${local.function_name}"
   description = "Coralogix Send Your Data key Secret"
 }
 
 resource "aws_secretsmanager_secret_version" "service_user" {
-  count         = var.layer_arn != "" && var.create_secret == "True"  ? 1 : 0
+  count         = var.secret_manager_enabled && var.create_secret == "True"  ? 1 : 0
   depends_on    = [aws_secretsmanager_secret.private_key_secret]
   secret_id     = aws_secretsmanager_secret.private_key_secret[count.index].id
   secret_string = var.private_key